I added compliance and a master service
-After compliance was made I populated it with semi-relevant information
The difficult part thats not done is the fact that the button still
doesn't want to move to the far right. We also need more helpful
information than we have right now.

-I made a master service so that if there is a value I need regardless
of the component it will be stored and inherted by what ever service
which will then in turn save that value. I used it to make the boolean,
if applications exist, since some buttons will have to fire differently
depending on the state of the user