Somewhat finished Sending, missing recieving
-See as it stands we are missing out on both cause we haven't provided
where the user is going to come from. This the same problem with
recieveing. How will the actual user input their name. I have Ideas, but
I'll rather see what yewo did first

-The back back-end of how messages are recieved are dependant on the
datacontext in the MessagesRepository and we will need to change that
and simply and a respective method in MessageController, just like how
we do when we need something in AccountController