I finished MessagesController
It was quiet difficult cause he kept including difficult and unasked for
methods that, to be honest will be quiet necessary. See we are using
firebase and he is using sql, so I had to find work arounds each time a
call or reference to a dbcontext was made. Then I found out that he had
set up previously alot of classes and extensions that I didn't yet have
access to. But once there I feel like it should work alright. At least
it compiles.

The problem is this. I have only gone half way and I'm not sure if my
changes will have consequences, but we will see.