Styling and functionality
-[UPDATE] Yeah, so it compiles yeah, but all you get is whitescreen.
I'll have to deal with this later doe`
-I won't lie, alot of this didn't seem necessary, but I figured it was
good for the user experience. I left comments everywhere where I felt
things that aren't really necessary can be removed, but feel free to
change where ever else you believe necessary. See the major problem with
all of this is that it takes OOB to heart. There is too many files and
too many injections for me to be comfortable. Some real refactoring
should be done here if we have the time

-I installed a depencency called microsoft/signalr. Run npm i
@microsoft/signalr to install it. Although I'm scared I missed a few
steps to properly configure it, since it compiled without any problems